+++ api/poll/v1/poll.go
@@ -25,7 +25,7 @@ type PollReq struct {
 	g.Meta       `path:"/poll" tags:"Poll" method:"put" summary:"投票"`
 	Method       consts.PollMethod `json:"method" dc:"投票方式" v:"required|enums"`
 	SentenceUUID string            `json:"sentence_uuid" dc:"句子 UUID" v:"required|length:36"`
-	MarkID       int64             `json:"mark_id" dc:"标记" v:"required-unless:method,1"`
+	MarkID       uint              `json:"mark_id" dc:"标记" v:"required-unless:method,1"`
 	Comment      string            `json:"comment" dc:"理由" v:"required-if:method,2,3|length:1,255"`
 }
